Fix server subclassing
