+++ Tella/Domain/Entity/GDrive/GDriveServer.swift
@@ -17,23 +17,14 @@ class GDriveServer: Server {
 
     init(id: Int? = nil,
          name: String? = "Google Drive",
-         serverURL: String? = nil,
-         username: String? = nil,
-         password: String? = nil,
-         accessToken: String? = nil,
          rootFolder: String,
-         serverType: ServerConnectionType? = .gDrive,
-         allowMultiple: Bool? = false
-    ) {
+         serverType: ServerConnectionType? = .gDrive) {
+        
         self.rootFolder = rootFolder
         super.init(id: id,
                    name: name,
-                   serverURL: serverURL,
-                   username: username,
-                   password: password,
-                   accessToken: accessToken,
-                   serverType: serverType
-        )
+                   serverType: serverType,
+                   allowMultipleConnections: false)
     }
 
     required init(from decoder: Decoder) throws {
@@ -41,7 +32,6 @@ class GDriveServer: Server {
         self.rootFolder = try container.decode(String?.self, forKey: .rootFolder)
         try super.init(from: decoder)
         self.serverType = .gDrive
-        self.allowMultiple = false
     }
 
     override func encode(to encoder: Encoder) throws {
