+++ Tella/Domain/Entity/Report/TellaServer.swift
@@ -1,0 +1,60 @@
+//  Tella
+//
+//  Copyright © 2022 INTERNEWS. All rights reserved.
+//
+
+import Foundation
+
+class TellaServer : WebServer {
+    
+    var activatedMetadata : Bool?
+    var backgroundUpload : Bool?
+    var projectId : String?
+    var slug : String?
+    var autoUpload: Bool?
+    var autoDelete: Bool?
+    
+    enum CodingKeys: String, CodingKey {
+        case id = "c_server_id"
+        case activatedMetadata = "c_activated_metadata"
+        case backgroundUpload = "c_background_upload"
+        case projectId = "c_api_project_id"
+        case slug = "c_slug"
+        case autoUpload = "c_auto_upload"
+        case autoDelete = "c_auto_delete"
+    }
+
+    init(id: Int? = nil,
+         name: String? = nil,
+         serverURL: String? = nil,
+         username: String? = nil,
+         password: String? = nil,
+         accessToken: String? = nil,
+         activatedMetadata: Bool? = nil,
+         backgroundUpload: Bool? = nil,
+         projectId: String? = nil,
+         slug: String? = nil,
+         autoUpload: Bool,
+         autoDelete: Bool,
+         serverType: ServerConnectionType? = .tella) {
+        
+        self.activatedMetadata = activatedMetadata
+        self.backgroundUpload = backgroundUpload
+        self.projectId = projectId
+        self.slug = slug
+        self.autoUpload = autoUpload
+        self.autoDelete = autoDelete
+        
+        super.init(id: id,
+                   name: name,
+                   serverURL: serverURL,
+                   username: username,
+                   password: password,
+                   accessToken: accessToken,
+                   serverType: serverType)
+    }
+    
+    required init(from decoder: Decoder) throws {
+        try super.init(from: decoder)
+    }
+}
